feat(orb-ui orb-qa): Add <tag> data-* attr/props to elements of interest
Add either static or dynamic labeling for elements ids and values in order
to make automated tests easier and more controllable.

* Add <data-orb-qa-id> attribute to some elements for testing

data-* attributes are reserved to be created in the dom to represent this sort of
meta info for accessibility of machines. Angular also enables these objects to be
bound dinamically, so we can have identification for each individual cell of a table,
for instance.